requirements.txt
python-telegram-bot>=20.0
instamojo>=1.2.0
qrcode>=7.4.2
Pillow>=10.0.0
APScheduler>=3.10.0
python-dotenv>=1.0.0
